#781 [REFACTOR] 검색 비동기 로딩 및 에러 핸들링을 Suspense, ErrorBoundary로 처리하도록 변경
