+++ src/components/AppBar/AppBar/AppBar.module.css
@@ -1,10 +1,10 @@
 .appBar {
   padding: 0.875rem 1.25rem;
+  width: 100%;
+  max-width: 600px;
   display: flex;
   justify-content: space-between;
   align-items: center;
-  width: 100%;
-  max-width: 600px;
   background-color: var(--white);
   position: fixed;
   top: 0;
