+++ src/components/List/List.module.css
@@ -1,0 +1,8 @@
+.list {
+  padding: 0.25rem var(--default-margin) 0 var(--default-margin);
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 0.5rem;
+  overflow-y: auto;
+}
