+++ src/components/Loading/FetchLoading/FetchLoading.module.css
@@ -1,7 +1,5 @@
 .loading {
   flex: 1;
-  width: 100%;
-  height: 100%;
   display: flex;
   justify-content: center;
   align-items: center;
