--- src/components/PostList/PostList.module.css
@@ -1,14 +1,0 @@
-.posts {
-  margin: 0 var(--default-margin);
-  padding-top: 4.25rem;
-  padding-bottom: 4rem;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  row-gap: 0.5rem;
-  font-size: 0.5rem;
-}
-
-.to {
-  width: 100%;
-}
