--- src/components/PostList/index.js
@@ -1,1 +1,0 @@
-export { default as PostList } from './PostList.jsx';
