+++ src/components/ServerErrorFallback/ServerErrorFallback.jsx
@@ -1,4 +1,4 @@
-import { MdOutlineReplay } from 'react-icons/md';
+import { ResetButton } from '@/components/ServerErrorFallback';
 
 import styles from './ServerErrorFallback.module.css';
 
@@ -8,8 +8,8 @@ export default function ServerErrorFallback({ reset }) {
       <div className={styles.errorMessage}>
         <span>잠시 후 다시 시도해주세요</span>
         <span>요청을 처리하는데 실패했어요</span>
+        <ResetButton reset={reset} />
       </div>
-      <MdOutlineReplay className={styles.retry} onClick={reset} />
     </div>
   );
 }
