+++ src/components/ServerErrorFallback/ServerErrorFallback.module.css
@@ -1,4 +1,5 @@
 .serverError {
+  flex: 1;
   margin: 0 var(--default-margin);
   padding: 2rem;
   aspect-ratio: 3.5;
@@ -26,16 +27,4 @@
 
 .errorMessage > span:nth-child(2) {
   font-size: 0.75rem;
-}
-
-.retry {
-  padding: 10px;
-  font-size: 20px;
-  transition: 400ms all ease;
-  border-radius: 100%;
-  cursor: pointer;
-}
-
-.retry:hover {
-  transform: rotateZ(-360deg);
 }
