+++ src/hooks/useSearch.jsx
@@ -1,7 +1,7 @@
 import { useSearchParams } from 'react-router-dom';
 
 import { searchByBoard } from '@/apis';
-import { usePagination } from '@/hooks';
+import { useSuspensePagination } from '@/hooks';
 import { QUERY_KEY, STALE_TIME } from '@/constants';
 
 export default function useSearch({ boardId }) {
@@ -9,27 +9,17 @@ export default function useSearch({ boardId }) {
   const params = Object.fromEntries(searchParams.entries());
   const paramsLength = Object.keys(params).length;
 
-  const { data, ref, isLoading, isFetching, status, isError, error, refetch } =
-    usePagination({
-      queryKey: [QUERY_KEY.search, boardId, params],
-      queryFn: ({ pageParam }) =>
-        searchByBoard({
-          boardId,
-          page: pageParam,
-          params,
-        }),
-      staleTime: STALE_TIME.searchList,
-      enabled: paramsLength > 0,
-    });
+  const { data, ref, isFetching, refetch } = useSuspensePagination({
+    queryKey: [QUERY_KEY.search, boardId, JSON.stringify(params)],
+    queryFn: ({ pageParam }) =>
+      searchByBoard({
+        boardId,
+        page: pageParam,
+        params,
+      }),
+    staleTime: STALE_TIME.searchList,
+    enabled: paramsLength > 0,
+  });
 
-  return {
-    data,
-    ref,
-    isLoading,
-    isFetching,
-    status,
-    isError,
-    error,
-    refetch,
-  };
+  return { data, ref, isFetching, refetch };
 }
