+++ src/pages/ExamReviewsPage/ExamReviewsPage.module.css
@@ -1,4 +1,5 @@
 .container {
+  padding-bottom: 5rem;
   width: 100%;
   height: 100vh;
   display: flex;
