--- src/pages/ExamReviewsPage/components/ExamReviewSearchList/ExamReviewSearchList.module.css
@@ -1,16 +1,0 @@
-.list {
-  padding: 0.25rem 1rem 0 1rem;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 0.5rem;
-  overflow: hidden;
-}
-
-.to {
-  width: 100%;
-}
-
-.loading {
-  padding-bottom: 7rem;
-}
