--- src/pages/ExamReviewsPage/components/ExamReviews/ExamReviewsErrorFallback.jsx
@@ -1,8 +1,0 @@
-import { ServerErrorFallback } from '@/components';
-
-export default function ExamReviewsErrorFallback({
-  error,
-  resetErrorBoundary,
-}) {
-  return <ServerErrorFallback reset={resetErrorBoundary} />;
-}
