+++ src/pages/ExamReviewsPage/index.js
@@ -1,7 +1,5 @@
 export { default as ExamReviewsPage } from './ExamReviewsPage';
 
-export { default as ExamReviews } from './components/ExamReviews/ExamReviews';
-export { default as ExamReviewsSuspense } from './components/ExamReviews/ExamReviewsSuspense';
-export { default as ExamReviewsErrorFallback } from './components/ExamReviews/ExamReviewsErrorFallback';
-
-export { default as ExamReviewSearchList } from './components/ExamReviewSearchList/ExamReviewSearchList';
+export { default as SearchExamReviews } from './components/SearchExamReviews/SearchExamReviews';
+export { default as SearchExamReviewsSuspense } from './components/SearchExamReviews/SearchExamReviewsSuspense';
+export { default as SearchExamReviewsErrorFallback } from './components/SearchExamReviews/SearchExamReviewsErrorFallback';
