--- src/pages/PostPage/PostSearchPage/PostSearchPage.module.css
@@ -1,9 +1,0 @@
-.container {
-  width: 100%;
-  height: 100vh;
-  overflow-y: auto;
-}
-
-.to {
-  width: 100%;
-}
