--- src/pages/PostPage/PostSearchPage/index.js
@@ -1,1 +1,0 @@
-export { default as PostSearchPage } from './PostSearchPage';
