+++ src/pages/PostPage/index.js
@@ -1,5 +1,4 @@
 export { default as PostPage } from './PostPage/PostPage';
 export { default as PostContent } from './PostPage/PostContent';
-export { default as PostSearchPage } from './PostSearchPage/PostSearchPage';
 export { default as PostWritePage } from './PostWritePage/PostWritePage';
 export { default as PostEditPage } from './PostEditPage/PostEditPage';
