+++ src/pages/PostsPage/PostsPage.module.css
@@ -1,11 +1,10 @@
 .container {
+  padding-top: 3.375rem;
   width: 100%;
   height: 100vh;
   display: flex;
   flex-direction: column;
-  align-items: center;
   font-size: 0.5rem;
-  padding-top: 3.375rem;
   overflow-y: auto;
 }
 
@@ -14,67 +13,18 @@
   padding: 0.625rem var(--default-margin);
 }
 
-.notification_bar {
-  width: 100%;
+.notificationBar {
   padding: 0.688rem 0.938rem;
+  width: 100%;
   display: flex;
   align-items: center;
   gap: 0.861rem;
   border-radius: 0.3125rem;
   background-color: var(--pink-1);
-  /* font-size: 0.688rem; */
   font-size: 0.8125rem;
   color: var(--pink-3);
   box-sizing: border-box;
   cursor: pointer;
-}
-
-.side_menu_btn {
-  width: 1.25rem;
-  height: 0.9375rem;
-  display: flex;
-}
-
-.notification_icon {
-  width: 0.663rem;
-  height: 0.791rem;
-  display: flex;
-}
-
-.to {
-  width: 100%;
-}
-
-.keyword_box {
-  display: flex;
-  width: 100%;
-  justify-content: baseline;
-}
-
-.keyword {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding: 0.1875rem 0.875rem;
-  border-radius: 1.25rem;
-  background: var(--blue-0);
-  margin-right: 0.3125rem;
-  color: var(--blue-3);
-
-  p {
-    font-size: 0.6875rem;
-  }
-}
-
-.content {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  font-size: 0.5rem;
-  /* 양옆 여백 */
-  padding: 0.25rem 1rem 12rem 1rem;
-  row-gap: 0.5rem;
 }
 
 .writeButton {
