+++ src/pages/SearchPage/SearchPage.module.css
@@ -1,0 +1,12 @@
+.container {
+  padding-top: 4rem;
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  overflow-y: auto;
+}
+
+.to {
+  width: 100%;
+}
