+++ src/pages/SearchPage/index.js
@@ -1,0 +1,5 @@
+export { default as SearchPage } from './SearchPage';
+
+export { default as SearchResults } from './components/SearchResults/SearchResults';
+export { default as SearchResultsSuspense } from './components/SearchResults/SearchResultsSuspense';
+export { default as SearchResultsErrorFallback } from './components/SearchResults/SearchResultsErrorFallback';
